+++ Src/Gtk2/DasherControl.cpp
@@ -319,7 +319,7 @@ gint CDasherControl::KeyPressEvent(GdkEventKey *event) {
 }
 
 void CDasherControl::SliderEvent() {
-  SetLongParameter(LP_MAX_BITRATE, GTK_RANGE(m_pSpeedHScale)->adjustment->value * 100);
+  SetLongParameter(LP_MAX_BITRATE, long(GTK_RANGE(m_pSpeedHScale)->adjustment->value * 100));
 }
 
 void CDasherControl::CanvasDestroyEvent() {
